DevLog 2/15
Added interface for objects that deal damage so that adding hovering
damage text is simpler in the future. Then added damage text for
everything. Fixed bugs with stat buttons. Added mod and inventory
buttons to begin work on the mod and inventory system.

+++ Assets/Scripts/Old Scripts/Enemies.cs
@@ -54,7 +54,7 @@ public abstract class Enemies : MonoBehaviour, IDamageable
         }
     }
 
-    public void TakeDamage(float damage)
+    public void TakeDamage(int damage)
     {
         health -= (int)damage;
     }
